decoder: change scaler values to long, fixed bug in timestamp decoding

+++ common-tools/clas-detector/src/main/java/org/jlab/detector/decode/CLASDecoder.java
@@ -337,7 +337,7 @@ public class CLASDecoder {
             scalerBANK.setShort("channel", i, (short) scalerDGTZ.get(i).getDescriptor().getChannel());
             scalerBANK.setByte("helicity", i, (byte) scalerDGTZ.get(i).getSCALERData(0).getHelicity());
             scalerBANK.setByte("quartet", i, (byte) scalerDGTZ.get(i).getSCALERData(0).getQuartet());
-            scalerBANK.setInt("value", i, scalerDGTZ.get(i).getSCALERData(0).getValue());
+            scalerBANK.setLong("value", i, scalerDGTZ.get(i).getSCALERData(0).getValue());
         }
 //        if(scalerBANK.rows()>0)scalerBANK.show();
         return scalerBANK;
